Most modules are now self-contained, using the anchor pattern.
Ensure that using the ordering operators on modules (i.e. classes with
subclasses), as in "Class['custom_repo'] -> Class['mymodule']", results in
custom_repo being declared/applied and finished *before* anything gets done for
mymodule.  This means a module written in this way is indeed self-contained.
This is the behaviour one would naively expect from Puppet, but in reality it
doesn't work that way.  See issue #8040
(http://projects.puppetlabs.com/issues/8040). A worthwhile discussion on the
subject can be found here:
https://groups.google.com/d/topic/puppet-users/6SdcdBtye9A/discussion

There are still some modules in our tree that need refactoring in this regard,
especially more complex modules like opsview and nagios.