Proper, LSB-compliant opsview-agent init script
This init script is more robust and compliant than the original,
supplied by Opsview.com. It will not work in RHEL/CentOS without slight
porting, as the log functions log_daemon_msg and log_end_msg don't exist
there. To make this script completely compatible it should be refactored
to not use any log_*_msgs at all. This is not a priority yet.